Separated out the InfoUI from the BasicSidebar
